+++ src/models/index.ts
@@ -12,6 +12,7 @@ export * from './confidential-storage';
 export * from './consent-rule';
 export * from './consent-access';
 export * from './crypto';
+export * from './dataspace-discovery';
 export * from './device-license';
 export * from './did';
 export * from './fhir-documents';